Allow empty arrays to be returned
In Jsonata it is a valid response return `[]any{}` from a transform, but
this was previously filtered out.

For example, in jsonata-js, the following input and transform result in
a response of `[]`.
Input:
```
{
  "Account": {
    "Account Name": "Firefly",
    "Order": [
    ]
  }
}
```
Transform: "Account.Order"

The same data and transform resulted in a "no results found" error in
jsonata-go.